Added initial files.
This commit reflects the original proportions generated. These were created by testing "If a fungi exists at 12, will it exist at 12 next year?" After some empirical analysis of our results, we see the Penicillium and Pezicula (which make up a good proportion of a tree community) are severely outcompeted. This commit is to save the original calculations before I try and alter the system.